fix: Tooltip positioning Sleek theme
The Sleek theme (and possibly others) use `absolute` positioning for the middle column. This causes the `.position()` call to return a value relative to the column origin for buttons inside the column. This causes the tooltip to often overlay the container that triggered it.

Using the document-relative `.offset()` works for both relative and absolute positioned columns.