feat: Allow resources to be cached
I feel like this was someone's intention all along. I just tied all the pieces together. When a new release is published `yarn run build` should be used to bump the revision. Then that is used to construct the cache-breaking URL query parameter. This way the client only need to pull the resources once for every version of KG.