Integrate Changes from Default Branch into Alpha  (#84)
* make the % for the heatCompression be %

1000 years was 100 times weaker than intended with heatCompression formula. ooups

* A few bugfixes & Easter eggs (#45)

* Prevent NULL year & season from showing in the log if you don't have Calendar

* Fix hash level not being properly reset at 0 hashes

* Easter egg will disappear on April 2

* Easter egg: pumpkin spice in October

* Hide redeemGiftBtn if elderBox resource is hidden from resource pane

* fix a dojo subscription leak from chiral (#58)

I noticed my tab running sluggishly, dumped the heap, and quickly
noticed about 64MB of retained closures in the ui/update topic
listeners. I noticed they had the `setCommand` function and found this
React component. It looks like there was a typo which prevents the
unsubscription. I was able to confirm with the debugger that
`componentDidMount` is called fairly frequently and that
`componentWillUmount` does not get called.

The best I could do to verify this fix, since I don't have a user login
server, was to change the condition to render the furthest ring tab and
then check the listeners via `dojo._topics["ui/update"]._listeners`.
Before the fix, this would grow by 1 each time you click in and out of
the space tab. After apply the fix, it did not grow, though it appears
`dojo.unsubscribe` simply deletes from the array, leaving it sparse. It
appears to run just fine.

* fix redshift applying calendar updates twice (#60)

The bug is that when applying redshift under certain conditions,
`game.calendar.fastForward()` is called twice. This is triggered when:

1. The `QUEUE_REDSHIFT` flag is enabled (true by default)
2. The queue is empty

The short-circuiting path must call `this.applyRedshift(x, true)` to
ignore the calendar, since `game.calendar.fastForward(x)` is called on
the other side of the `while` loop. Without passing `true`, the calendar
is advanced twice, meaning you get twice as many events, antimatter,
beacon relics, void, and years as you would normally. Originally
introduced in 66d451193.

You can reproduce this by pausing long enough for any of those values to
tick (e.g., a few seconds for relics) and noting the difference in
values compared to the number of days gained. With the patch, you no
longer get twice as many.

I chose to rename `daysOffset` to `daysOffsetLeft` even though they're
both the same value since it matches the other calls to `applyRedshift`
in this branch and maybe somewhat indicates that we're only doing part
of the work.

* Remove console spam from unused pacts gives karma (#65)

* Bringing RN requirement in line with description and Dead Space

* feat: Remove Kiwi IRC client (#72)

As far as I know, the integration has been broken for some time and the resources being loaded just slow down the page.

* Tame infinite prices (#76)

* Tame infinite prices

* Tame really, really large--but still finite--toDisplaySeconds

* Update expected outputs from toDisplaySeconds to match new behavior

* feat: Clean up BitBucket references

Also updates the manifest, thus a feature-level change.

* fix: Remove reference to non-existent source map

The reference to the source map causes a 404 every time devtools are being used.

* fix: Remove leftover kiwii reference (#77)

* fix: Remove obsolete Kongregate entrypoint (#78)

This file was used as an entrypoint to Kittens Game when it was accessed on Kongregate. This service is defunct today, and the file is no longer relevant.

* Minor QoL improvements for KGNet integration (#80)

* fix: Only show instructions if no saves exist

* feat: Auto-sync saves if login is valid on startup

* Log Bugfixes & SETI QoL (#79)

* Fix bugs related to recursive nature of renderConsoleLog

* Upon purchasing SETI, hide the log filter for astro events & auto-complete the ongoing astro event if there is one.

---------

Co-authored-by: AZiggurat <149694649+ZigguratV@users.noreply.github.com>
Co-authored-by: Brent-Call <90073067+Brent-Call@users.noreply.github.com>
Co-authored-by: Doug Patti <douglas.patti@gmail.com>
Co-authored-by: PG-Fuury <32613315+PG-Fuury@users.noreply.github.com>