Small batch of fixes (#135)
* fix: Invalid named-color `darkYellow`

The color is not a valid HTML named color. I replaced it with `olive`, which is a dark yellow (`#808000`).

* fix: Charon mission unlocks itself

Potentially a copy/paste mistake. There should be no need for a mission to unlock itself.

* fix: Invalid cat colors

Some of the named colors used for cats are not actual HTML named colors. I replaced them with colors from the cat palette.

* feat: Additional cat names

Those cute little fur balls...